+++ roqoqo/build.rs
@@ -131,7 +131,7 @@ impl<'ast> Visit<'ast> for Visitor {
     fn visit_item_struct(&mut self, i: &'ast ItemStruct) {
         // Check attributes
         for att in i.attrs.clone() {
-            let path = att.path.get_ident().map(|id| id.to_string());
+            let path = att.path().get_ident().map(|id| id.to_string());
             // only consider the derive attribute, if no derive attribute is present don't add anything
             // to the internal storage of the visitor
             if matches!(att.style, AttrStyle::Outer) && path == Some("derive".to_string()) {