+++ roqoqo/tests/integration/operations/multi_qubit_gate_operations.rs
@@ -451,7 +451,7 @@ fn test_matrix_output_three_multi_qubit_zz(qubits: Vec<usize>) {
 }
 
 #[test]
-fn test_clone_partial_eq_mulit_qubit_ms() {
+fn test_clone_partial_eq_multi_qubit_ms() {
     let qubits = vec![0, 1, 2];
 
     let gate = MultiQubitZZ::new(qubits.clone(), CalculatorFloat::FRAC_PI_2);
@@ -475,7 +475,7 @@ fn test_clone_partial_eq_mulit_qubit_ms() {
 }
 
 #[test]
-fn test_operate_mulit_qubit_zz() {
+fn test_operate_multi_qubit_zz() {
     let qubits = vec![0, 1, 2];
     let gate = MultiQubitZZ::new(qubits.clone(), CalculatorFloat::FRAC_PI_2);
     assert_eq!(gate.hqslang(), "MultiQubitZZ");